Add files via upload
Files contained here work here to create an operating system simulator. The simulator Incorporates the real world (i.e., advanced academia and/or industry) conditions of managing a larger scale project and provides the quality experience of the design and operational decisions made by persons developing an operating system. This project was done in 4 parts. Unfortunately, in March COVID-19 restricted my access to additional assistance from my professors as they were trying to help everyone else affected by the virus as well. Therefore, the system operates correctly up to part 4. Currently the system is not finished fully and will not run because there are implementations from part 04 that were never resolved, mainly in regards to pre-emptive scheduling and memory complications.